+++ compiler/commands.nim
@@ -623,6 +623,9 @@ proc processSwitch*(switch, arg: string, pass: TCmdLinePass, info: TLineInfo;
   of "outdir":
     expectArg(conf, switch, arg, pass, info)
     conf.outDir = processPath(conf, arg, info, notRelativeToProj=true)
+  of "depfile":
+    expectArg(conf, switch, arg, pass, info)
+    conf.depfile = processPath(conf, arg, info, notRelativeToProj=true).AbsoluteFile
   of "usenimcache":
     processOnOffSwitchG(conf, {optUseNimcache}, arg, pass, info)
   of "docseesrcurl":